backend/requirements.txt
Django==5.0.14
djangorestframework
djangorestframework-simplejwt
django-filter
mysqlclient
